+++ MK4duo/src/core/commands/gcode/feature/g61.h
@@ -41,7 +41,7 @@ inline void gcode_G61(void) {
   const uint8_t slot = parser.byteval('S');
 
   if (slot >= NUM_POSITON_SLOTS) {
-    SERIAL_LMV(ER, MSG_INVALID_POS_SLOT, (int)NUM_POSITON_SLOTS);
+    SERIAL_LMV(ER, MSG_INVALID_POS_SLOT, NUM_POSITON_SLOTS);
     return;
   }
